requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2-binary==2.8.5
# psycopg2

